pyproject.toml
[project]
name = "obsidian-bot"
version = "0.1.0"
description = "Telegram bot for capturing notes into an Obsidian vault"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=22.6",
    "httpx>=0.28",
    "readability-lxml>=0.8",
    "markdownify>=0.14",
    "google-genai>=1.0",
]

[project.scripts]
obsidian-bot = "obsidian_bot.main:main"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.12.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"
